🐛 fix: reduce the maximum time for the cache to remain validated for (#762) - Reduced the maximum time to 256 seconds that is 4 minutes and 15 seconds.

+++ src/parser.rs
@@ -57,7 +57,7 @@ pub struct Config {
     pub operating_system_tls_certificates: bool,
     /// It stores the http-caching's validity time that is how long the http cache will remain
     /// valid for all the resources that was loaded on the page.
-    pub http_cache_expiry_time: u16,
+    pub http_cache_expiry_time: u8,
 }
 
 impl Config {
